// plasma-ui/build/openapi/model/entityTypeDTO.ts
/**
 * PLASMA Data Modeling Application
 * Service for modeling data sources
 *
 * OpenAPI spec version: 0.1.3
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EntityConceptDTO } from './entityConceptDTO';

export interface EntityTypeDTO {
    x?: number;
    y?: number;
    id?: string;
    uuid?: string;
    label?: string;
    originalLabel?: string;
    description?: string;
    concept?: EntityConceptDTO;
    mappedToData?: boolean;
}