// plasma-ui/build/openapi/model/schemaNodeDTO.ts
/**
 * PLASMA Data Modeling Application
 * Service for modeling data sources
 *
 * OpenAPI spec version: 0.1.3
 * Contact: pomp@uni-wuppertal.de
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SyntacticOperationDTO } from './syntacticOperationDTO';

export interface SchemaNodeDTO {
    x?: number;
    y?: number;
    id?: string;
    label?: string;
    operations?: Array<SyntacticOperationDTO>;
    isValid?: boolean;
    _class: string;
}